// admcentro-web-main/src/interfaces/Ieventualities.ts
// Generated by https://quicktype.io

import { Contract } from './Icontracts'
import { Iproperty } from './Iproperties'

export interface IEventualitiesResponse {
  results: number
  code: number
  status: string
  ok: boolean
  data: IEventuality[]
}

export interface IEventuality {
  id: number
  ContractId: number | null
  PropertyId: number | null
  clientAmount: number
  ownerAmount: number
  clientPaid: boolean
  ownerPaid: boolean
  description: string
  expiredDate: string
  createdAt: string
  updatedAt: string
  deletedAt: null
  Contract: Contract | null
  Property: Iproperty | null
}